# backend/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TransactionDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"projectId": {
						"type": "number",
						"format": "double"
					},
					"accountId": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"credited": {
						"type": "boolean"
					},
					"modifiedBy": {
						"type": "string"
					},
					"intrestRate": {
						"type": "number",
						"format": "double"
					},
					"transactionDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"details",
					"projectId",
					"accountId",
					"amount",
					"credited",
					"intrestRate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResourceDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"location": {
						"type": "string"
					},
					"sourceId": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"group": {
						"type": "string"
					}
				},
				"required": [
					"location",
					"type",
					"group"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"regId": {
						"type": "string"
					},
					"ownerId": {
						"type": "string"
					},
					"caId": {
						"type": "string"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"name",
					"address",
					"country",
					"regId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectRespDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"countryName": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"actualMaturityDate": {
						"type": "string",
						"format": "date-time"
					},
					"overallCost": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"ownerName": {
						"type": "string"
					},
					"legalId": {
						"type": "string"
					},
					"maturityLockingPeriod": {
						"type": "number",
						"format": "double"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"entity": {
						"$ref": "#/components/schemas/EntityDTO"
					},
					"resources": {
						"items": {
							"$ref": "#/components/schemas/ResourceDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"address",
					"countryName",
					"startDate",
					"entity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InvestmentDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"investmentType": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"accountId": {
						"type": "number",
						"format": "double"
					},
					"modifiedBy": {
						"type": "string"
					}
				},
				"required": [
					"investmentType",
					"amount",
					"accountId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccountRespDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"investorId": {
						"type": "number",
						"format": "double"
					},
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/TransactionDTO"
						},
						"type": "array"
					},
					"investments": {
						"items": {
							"$ref": "#/components/schemas/InvestmentDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"currency",
					"investorId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvestorRespDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nomineeDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"emergencyContact": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"personalDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"ownerId": {
						"type": "string"
					},
					"caId": {
						"type": "string"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"accounts": {
						"items": {
							"$ref": "#/components/schemas/AccountRespDTO"
						},
						"type": "array"
					},
					"projects": {
						"items": {
							"$ref": "#/components/schemas/ProjectRespDTO"
						},
						"type": "array"
					},
					"resources": {
						"items": {
							"$ref": "#/components/schemas/ResourceDTO"
						},
						"type": "array"
					},
					"documents": {
						"items": {
							"properties": {
								"status": {
									"type": "boolean"
								},
								"docUrl": {
									"type": "string"
								},
								"docName": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"status",
								"docUrl",
								"docName",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"emergencyContact",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvestorReqDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nomineeDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"emergencyContact": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"personalDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"ownerId": {
						"type": "string"
					},
					"caId": {
						"type": "string"
					},
					"projectIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"resourceIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"documents": {
						"items": {
							"properties": {
								"status": {
									"type": "boolean"
								},
								"docUrl": {
									"type": "string"
								},
								"docName": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"status",
								"docUrl",
								"docName",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"emergencyContact",
					"userId",
					"projectIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial__docName-string--docUrl-string--status-boolean__": {
				"properties": {
					"docName": {
						"type": "string"
					},
					"docUrl": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"ContentDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ContentDTO_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"AccountReqDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"investorId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"currency",
					"investorId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nexgen-services",
		"version": "1.0.0",
		"description": "**Edit a file, create a new file, and clone from Bitbucket in under 2 minutes**",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/investors": {
			"get": {
				"operationId": "GetAllInvestors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InvestorRespDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateInvestor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InvestorReqDTO"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateInvestor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InvestorReqDTO"
							}
						}
					}
				}
			}
		},
		"/investors/{id}": {
			"get": {
				"operationId": "GetInvestorById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/investors/{id}/nomineeDetails": {
			"put": {
				"operationId": "UpdateInvestorNomineeDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {},
								"additionalProperties": {},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/investors/{id}/emergencyContact": {
			"put": {
				"operationId": "UpdateEmergencyContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {},
								"additionalProperties": {},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/investors/{id}/personalDetails": {
			"put": {
				"operationId": "UpdateInvestorPersonalDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {},
								"additionalProperties": {},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/investors/{id}/projects/assign": {
			"put": {
				"operationId": "AssignProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/investors/{id}/documents": {
			"put": {
				"operationId": "AddDocumentDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "boolean"
									},
									"docUrl": {
										"type": "string"
									},
									"docName": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"status",
									"docUrl",
									"docName"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetDocumentsByInvestorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"docUrl": {
												"type": "string"
											},
											"docName": {
												"type": "string"
											}
										},
										"required": [
											"docUrl",
											"docName"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/investors/{id}/documents/{documentId}": {
			"put": {
				"operationId": "UpdateDocumentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial__docName-string--docUrl-string--status-boolean__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDocumentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/InvestorRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/contents": {
			"get": {
				"operationId": "GetAllContents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ContentDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateContent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContentDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContentDTO"
							}
						}
					}
				}
			}
		},
		"/contents/{id}": {
			"get": {
				"operationId": "GetContentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ContentDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteContent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateContent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ContentDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ContentDTO_"
							}
						}
					}
				}
			}
		},
...
		"/accounts": {
			"get": {
				"operationId": "GetAllAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AccountRespDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/AccountRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountReqDTO"
							}
						}
					}
				}
			}
		},
		"/accounts/{id}": {
			"get": {
				"operationId": "GetAccountById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/AccountRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/accounts/investor/{investorId}": {
			"get": {
				"operationId": "GetAccountsByInvestorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AccountRespDTO"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "investorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/accounts/{id}/remove-investor": {
			"patch": {
				"operationId": "RemoveInvestorFromAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/AccountRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}